new crc system. Collisions

+++ game/src/sprite/spriteCollision.hpp
@@ -5,7 +5,7 @@ class genericSprite;
 class spriteCollision
 {
 public:
-	typedef std::map<int, boost::shared_ptr<genericSprite> > spriteContainer;
+	typedef std::map<unsigned short, boost::shared_ptr<genericSprite> > spriteContainer;
 	spriteCollision(spriteContainer &Sprites);
 	~spriteCollision();
 	void update(int spriteID);
@@ -15,7 +15,7 @@ public:
 	{
 		float x, y;
 	};
-	typedef std::map<int, boost::shared_ptr<collisionInfo> > collisionInfoContainer;
+	typedef std::map<unsigned short, boost::shared_ptr<collisionInfo> > collisionInfoContainer;
 	collisionInfoContainer infos;
 	spriteContainer &Sprites;
